Legg typer og mappingfunksjoner for vilkårsoppdatering i egen fil
